Codigo/DeliveryApi/requirements.txt
alembic==1.5.4
analytics-python==1.2.9
aniso8601==9.0.1
apispec==4.4.0
apispec-webframeworks==0.5.2
APScheduler==3.7.0
astroid==2.5.2
autopep8==1.5.5
boto3==1.16.56
botocore==1.19.63
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
cognitojwt==1.2.2
colorama==0.4.4
decorator==5.0.9
ecdsa==0.14.1
Flask==1.1.1
flask-apispec==0.11.0
Flask-APScheduler==1.11.0
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.9.1
mccabe==0.6.1
memoization==0.3.2
multi-key-dict==2.0.3
pep8==1.7.1
pika==1.2.0
PikaBus==1.3.6
platformdirs==2.4.0
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.7.4
pylint-flask-sqlalchemy==0.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2021.1
PyYAML==5.4.1
rabbitmq-pika-flask==1.2.10
requests==2.25.1
responses==0.13.3
retry==0.9.2
rsa==4.7.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
sqlalchemy-views==0.2.4
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.3
webargs==8.0.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
